refactor(repo): rename trace-propagation package to tracing
Rename @supabase/trace-propagation to @supabase/tracing for consistency
with naming conventions across the codebase.

Changes:
- Rename packages/shared/trace-propagation to packages/shared/tracing
- Update package name from @supabase/trace-propagation to @supabase/tracing
- Update all imports in supabase-js package
- Update workspace configuration
- Rename test-trace-propagation to test-tracing
