+++ packages/shared/tracing/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@supabase/trace-propagation",
+  "name": "@supabase/tracing",
   "version": "0.0.0-automated",
   "description": "W3C/OpenTelemetry trace context propagation utilities for Supabase JS SDK",
   "main": "dist/main/index.js",
@@ -16,7 +16,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/supabase/supabase-js.git",
-    "directory": "packages/shared/trace-propagation"
+    "directory": "packages/shared/tracing"
   },
   "keywords": [
     "tracing",
@@ -35,7 +35,7 @@
   "bugs": {
     "url": "https://github.com/supabase/supabase-js/issues"
   },
-  "homepage": "https://github.com/supabase/supabase-js/tree/master/packages/shared/trace-propagation",
+  "homepage": "https://github.com/supabase/supabase-js/tree/master/packages/shared/tracing",
   "dependencies": {
     "tslib": "2.8.1"
   },
