+++ packages/shared/tracing/src/defaults.ts
