+++ packages/shared/tracing/src/validate.ts
