+++ packages/shared/tracing/test/defaults.test.ts
