+++ packages/shared/tracing/tsconfig.json
