+++ packages/shared/tracing/tsconfig.module.json
