+++ test-tracing/supabase/.gitignore
