+++ test-tracing/supabase/config.toml
@@ -2,7 +2,7 @@
 # https://supabase.com/docs/guides/local-development/cli/config
 # A string used to distinguish different Supabase projects on the same host. Defaults to the
 # working directory name when running `supabase init`.
-project_id = "test-trace-propagation"
+project_id = "test-tracing"
 
 [api]
 enabled = true
