+++ test-tracing/test-trace-simple.js
